requirements.txt
Flask
flask-session
python-dotenv
pandas
PyMuPDF
